:recycle: Define VersionTag extension for Repository
