+++ src/commits.rs
@@ -1,5 +1,5 @@
 use crate::conventional_commit::ConventionalCommit;
-use crate::version_tag::get_latest_version_tag;
+pub use crate::version_tag::RepositoryVersionTagExtension;
 use git2::Oid;
 use git2::Repository;
 use std::error::Error;
@@ -13,7 +13,7 @@ use std::error::Error;
 pub fn fetch_commits_since_last_version(
     repository: &Repository,
 ) -> Result<Vec<ConventionalCommit>, Box<dyn Error>> {
-    match get_latest_version_tag(repository)? {
+    match repository.get_latest_version_tag()? {
         Some(version_tag) => fetch_commits_until(repository, version_tag.commit_oid),
         None => fetch_all_commits(repository),
     }
@@ -43,119 +43,6 @@ fn general_fetch_commits_until(
         .filter_map(|oid| repository.find_commit(oid).ok())
         .map(|commit| ConventionalCommit::from_git2_commit(commit))
         .collect())
-}
-
-#[cfg(test)]
-mod get_latest_version_tag_tests {
-    use crate::test_util::repo_init;
-    pub use crate::test_util::RepositoryTestExtensions;
-    use crate::version_tag::get_latest_version_tag;
-    use semver::Version;
-
-    #[test]
-    fn repository_does_not_have_tags() {
-        // Given
-        let (_temp_dir, repository) = repo_init(None);
-
-        // When
-        let result = get_latest_version_tag(&repository).unwrap();
-
-        // Then
-        assert!(result.is_none(), "Expected None, but got Some")
-    }
-
-    #[test]
-    fn repository_does_not_have_version_tags() {
-        // Given
-        let (_temp_dir, repository) = repo_init(Some(vec![":tada: initial release"]));
-        let commit = repository.find_commit_by_message(":tada: initial release");
-        repository.add_tag(commit.unwrap(), "tag_1");
-
-        // When
-        let result = get_latest_version_tag(&repository).unwrap();
-
-        // Then
-        assert!(result.is_none(), "Expected None, but got Some")
-    }
-
-    #[test]
-    fn repository_has_one_annotated_version_tag() {
-        // Given
-        let commit_message = ":tada: initial release";
-        let (_temp_dir, repository) = repo_init(Some(vec![commit_message]));
-        let commit = repository.find_commit_by_message(commit_message);
-        repository.add_tag(commit.unwrap(), "v1.0.0");
-
-        // When
-        let result = get_latest_version_tag(&repository).unwrap().unwrap();
-
-        // Then
-        assert_eq!(result.version, Version::parse("1.0.0").unwrap());
-        assert_eq!(
-            result.commit_oid,
-            repository
-                .find_commit_by_message(commit_message)
-                .unwrap()
-                .id(),
-            "Object IDs don't match"
-        );
-    }
-
-    #[test]
-    fn repository_has_one_not_annotated_version_tag() {
-        // Given
-        let commit_message = ":tada: initial release";
-        let (_temp_dir, repository) = repo_init(Some(vec![commit_message]));
-        let commit = repository.find_commit_by_message(commit_message).unwrap();
-        repository
-            .tag_lightweight("v1.0.0", commit.as_object(), false)
-            .unwrap();
-
-        // When
-        let result = get_latest_version_tag(&repository).unwrap().unwrap();
-
-        // Then
-        assert_eq!(result.version, Version::parse("1.0.0").unwrap());
-        assert_eq!(
-            result.commit_oid,
-            repository
-                .find_commit_by_message(commit_message)
-                .unwrap()
-                .id(),
-            "Object IDs don't match"
-        );
-    }
-
-    #[test]
-    fn repository_have_multiple_version_tags() {
-        // Given
-        let commit_messages = vec![
-            ":tada: initial release",
-            ":sparkles: new feature",
-            ":boom: everything is broken",
-        ];
-        let (_temp_dir, repository) = repo_init(Some(commit_messages.clone()));
-        let tags = vec!["v1.0.0", "v1.1.0", "v2.0.0"];
-        commit_messages
-            .iter()
-            .map(|commit| repository.find_commit_by_message(commit).unwrap())
-            .zip(tags)
-            .for_each(|(commit_id, tag)| repository.add_tag(commit_id, &tag));
-
-        // When
-        let result = get_latest_version_tag(&repository).unwrap().unwrap();
-
-        // Then
-        assert_eq!(result.version, Version::parse("2.0.0").unwrap());
-        assert_eq!(
-            result.commit_oid,
-            repository
-                .find_commit_by_message(commit_messages.last().unwrap())
-                .unwrap()
-                .id(),
-            "Object IDs don't match"
-        );
-    }
 }
 
 #[cfg(test)]
