+++ src/version_tag.rs
@@ -3,35 +3,39 @@ use regex::Regex;
 use semver::Version;
 use std::error::Error;
 
-/// Get the latest version tag.
-/// ## Returns
-/// [`VersionTag`] containing the latest version tag.
-pub fn get_latest_version_tag(
-    repository: &Repository,
-) -> Result<Option<VersionTag>, Box<dyn Error>> {
-    let mut version_tags: Vec<VersionTag> = Vec::new();
+pub trait RepositoryVersionTagExtension {
+    fn get_latest_version_tag(&self) -> Result<Option<VersionTag>, Box<dyn Error>>;
+}
 
-    let references = repository.references()?;
-    for reference in references {
-        let reference = reference?;
-        if reference.is_tag() {
-            if let Some(oid) = reference.target() {
-                let object = repository.find_object(oid, None)?;
+impl RepositoryVersionTagExtension for Repository {
+    /// Get the latest version tag.
+    /// ## Returns
+    /// [`VersionTag`] containing the latest version tag.
+    fn get_latest_version_tag(&self) -> Result<Option<VersionTag>, Box<dyn Error>> {
+        let mut version_tags: Vec<VersionTag> = Vec::new();
 
-                if let Ok(tag_object) = object.peel(ObjectType::Tag) {
-                    if let Some(tag) = tag_object.as_tag() {
-                        if let Some(version_tag) = VersionTag::from_annotated_tag(tag) {
-                            version_tags.push(version_tag);
+        let references = self.references()?;
+        for reference in references {
+            let reference = reference?;
+            if reference.is_tag() {
+                if let Some(oid) = reference.target() {
+                    let object = self.find_object(oid, None)?;
+
+                    if let Ok(tag_object) = object.peel(ObjectType::Tag) {
+                        if let Some(tag) = tag_object.as_tag() {
+                            if let Some(version_tag) = VersionTag::from_annotated_tag(tag) {
+                                version_tags.push(version_tag);
+                            }
                         }
+                    } else if let Some(version_tag) = VersionTag::from_lightweight_tag(reference) {
+                        version_tags.push(version_tag);
                     }
-                } else if let Some(version_tag) = VersionTag::from_lightweight_tag(reference) {
-                    version_tags.push(version_tag);
                 }
             }
         }
-    }
 
-    Ok(version_tags.iter().max().cloned())
+        Ok(version_tags.iter().max().cloned())
+    }
 }
 
 /// A structure that represent a version tag.
@@ -81,5 +85,118 @@ impl VersionTag {
     fn is_valid_version_tag(tag_name: &str) -> bool {
         let version_regex = Regex::new(r"^v\d+\.\d+\.\d+$").unwrap();
         version_regex.is_match(tag_name)
+    }
+}
+
+#[cfg(test)]
+mod get_latest_version_tag_tests {
+    use crate::test_util::repo_init;
+    pub use crate::test_util::RepositoryTestExtensions;
+    pub use crate::version_tag::RepositoryVersionTagExtension;
+    use semver::Version;
+
+    #[test]
+    fn repository_does_not_have_tags() {
+        // Given
+        let (_temp_dir, repository) = repo_init(None);
+
+        // When
+        let result = repository.get_latest_version_tag().unwrap();
+
+        // Then
+        assert!(result.is_none(), "Expected None, but got Some")
+    }
+
+    #[test]
+    fn repository_does_not_have_version_tags() {
+        // Given
+        let (_temp_dir, repository) = repo_init(Some(vec![":tada: initial release"]));
+        let commit = repository.find_commit_by_message(":tada: initial release");
+        repository.add_tag(commit.unwrap(), "tag_1");
+
+        // When
+        let result = repository.get_latest_version_tag().unwrap();
+
+        // Then
+        assert!(result.is_none(), "Expected None, but got Some")
+    }
+
+    #[test]
+    fn repository_has_one_annotated_version_tag() {
+        // Given
+        let commit_message = ":tada: initial release";
+        let (_temp_dir, repository) = repo_init(Some(vec![commit_message]));
+        let commit = repository.find_commit_by_message(commit_message);
+        repository.add_tag(commit.unwrap(), "v1.0.0");
+
+        // When
+        let result = repository.get_latest_version_tag().unwrap().unwrap();
+
+        // Then
+        assert_eq!(result.version, Version::parse("1.0.0").unwrap());
+        assert_eq!(
+            result.commit_oid,
+            repository
+                .find_commit_by_message(commit_message)
+                .unwrap()
+                .id(),
+            "Object IDs don't match"
+        );
+    }
+
+    #[test]
+    fn repository_has_one_not_annotated_version_tag() {
+        // Given
+        let commit_message = ":tada: initial release";
+        let (_temp_dir, repository) = repo_init(Some(vec![commit_message]));
+        let commit = repository.find_commit_by_message(commit_message).unwrap();
+        repository
+            .tag_lightweight("v1.0.0", commit.as_object(), false)
+            .unwrap();
+
+        // When
+        let result = repository.get_latest_version_tag().unwrap().unwrap();
+
+        // Then
+        assert_eq!(result.version, Version::parse("1.0.0").unwrap());
+        assert_eq!(
+            result.commit_oid,
+            repository
+                .find_commit_by_message(commit_message)
+                .unwrap()
+                .id(),
+            "Object IDs don't match"
+        );
+    }
+
+    #[test]
+    fn repository_have_multiple_version_tags() {
+        // Given
+        let commit_messages = vec![
+            ":tada: initial release",
+            ":sparkles: new feature",
+            ":boom: everything is broken",
+        ];
+        let (_temp_dir, repository) = repo_init(Some(commit_messages.clone()));
+        let tags = vec!["v1.0.0", "v1.1.0", "v2.0.0"];
+        commit_messages
+            .iter()
+            .map(|commit| repository.find_commit_by_message(commit).unwrap())
+            .zip(tags)
+            .for_each(|(commit_id, tag)| repository.add_tag(commit_id, &tag));
+
+        // When
+        let result = repository.get_latest_version_tag().unwrap().unwrap();
+
+        // Then
+        assert_eq!(result.version, Version::parse("2.0.0").unwrap());
+        assert_eq!(
+            result.commit_oid,
+            repository
+                .find_commit_by_message(commit_messages.last().unwrap())
+                .unwrap()
+                .id(),
+            "Object IDs don't match"
+        );
     }
 }
